% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdMTED-package.R
\docType{package}
\name{fdMTED-package}
\alias{fdMTED}
\alias{fdMTED-package}
\title{fdMTED: Tree edit distance for merge trees}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Jakob-Winkler/fdMTED}
  \item Report bugs at \url{https://github.com/Jakob-Winkler/fdMTED/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
