% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_edit_dist_helper.R
\name{TED}
\alias{TED}
\title{Split a string}
\usage{
TED(tree1, tree2)
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{What to split on.}
}
\value{
A character vector.
}
\description{
Split a string
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
